+++ Services/CustomerService.cs
@@ -1,5 +1,5 @@
 ﻿using Labb1_ASP.NET_API.Models;
-using Labb1_ASP.NET_API.Models.DTOs;
+using Labb1_ASP.NET_API.Models.DTOs.Customer;
 using Labb1_ASP.NET_API.Repositories.IRepositories;
 using Labb1_ASP.NET_API.Services.IServices;
 
@@ -28,6 +28,7 @@ namespace Labb1_ASP.NET_API.Services
         public async Task<CustomerWithIdDTO> GetCustomerByIdAsync(int id)
         {
             var existingCustomer = await _customerRepository.GetCustomerByIdAsync(id);
+            if(existingCustomer == null) { return null; }
             return new CustomerWithIdDTO
             {
                 Id = existingCustomer.Id,
@@ -40,6 +41,7 @@ namespace Labb1_ASP.NET_API.Services
         public async Task<CustomerWithIdDTO> GetCustomerByPhoneNumberAsync(string number)
         {
             var existingCustomer = await _customerRepository.GetCustomerByPhoneNumberAsync(number);
+            if (existingCustomer == null) { return null; }
             return new CustomerWithIdDTO
             {
                 Id = existingCustomer.Id,
@@ -48,6 +50,7 @@ namespace Labb1_ASP.NET_API.Services
                 PhoneNumber = existingCustomer.PhoneNumber,
             };
         }
+
         public async Task AddCustomerAsync(CustomerDTO customerDto)
         {
             await _customerRepository.AddCustomerAsync(new Customer
@@ -61,13 +64,24 @@ namespace Labb1_ASP.NET_API.Services
         public async Task EditCustomerAsync(CustomerDTO customerDto, int id)
         {
             var existingCustomer = await _customerRepository.GetCustomerByIdAsync(id);
-
+            if(existingCustomer == null)
+            {
+                throw new InvalidOperationException($"Customer with Id.{id} was not found!");
+            }
             existingCustomer.FirstName = customerDto.FirstName;
             existingCustomer.LastName = customerDto.LastName;
             existingCustomer.PhoneNumber = customerDto.PhoneNumber;
+
+            await _customerRepository.EditCustomerAsync(existingCustomer);
         }
+
         public async Task DeleteCustomerAsync(int id)
         {
+            var deleteCustomer = await _customerRepository.GetCustomerByIdAsync(id);
+            if (deleteCustomer == null)
+            {
+                throw new InvalidOperationException($"Customer with Id.{id} was not found!");
+            }
             await _customerRepository.DeleteCustomerAsync(id);
         }
     }
